backend/requirements.txt
fastapi
uvicorn
pydantic
deep-translator
translators
unbabel-comet